examples/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppt-tools"
version = "1.0.0"
description = "Extract and generate PowerPoint presentations from/to JSON data"
authors = [
    {name = "PPT Extractor", email = "example@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=1.0",
]

[project.scripts]
ppt-extractor = "ppt_extractor:main"
ppt-generator = "ppt_generator:main"

[project.urls]
Homepage = "https://github.com/example/ppt-extractor"
Issues = "https://github.com/example/ppt-extractor/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ppt_extractor*"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true